+++ src/getStepStartStates.ts
@@ -12,6 +12,7 @@ import { STEP_CLOUD_STORAGE_BUCKETS } from './steps/storage';
 import { STEP_API_SERVICES } from './steps/service-usage';
 import { deserializeIntegrationConfig } from './utils/integrationConfig';
 import { STEP_IAM_ROLES, STEP_IAM_SERVICE_ACCOUNTS } from './steps/iam';
+import { STEP_RESOURCE_MANAGER_IAM_POLICY } from './steps/resource-manager';
 
 async function getEnabledServiceNames(
   config: IntegrationConfig,
@@ -85,5 +86,8 @@ export default async function getStepStartStates(
     ),
     [STEP_IAM_ROLES]: createStepStartState(ServiceUsageName.IAM),
     [STEP_IAM_SERVICE_ACCOUNTS]: createStepStartState(ServiceUsageName.IAM),
+    [STEP_RESOURCE_MANAGER_IAM_POLICY]: createStepStartState(
+      ServiceUsageName.RESOURCE_MANAGER,
+    ),
   };
 }