+++ src/index.test.ts
@@ -22,6 +22,7 @@ import { STEP_CLOUD_STORAGE_BUCKETS } from './steps/storage';
 import { STEP_API_SERVICES } from './steps/service-usage';
 import { parseServiceAccountKeyFile } from './utils/parseServiceAccountKeyFile';
 import { STEP_IAM_ROLES, STEP_IAM_SERVICE_ACCOUNTS } from './steps/iam';
+import { STEP_RESOURCE_MANAGER_IAM_POLICY } from './steps/resource-manager';
 
 interface ValidateInvocationInvalidConfigTestParams {
   instanceConfig?: Partial<IntegrationConfig>;
@@ -96,6 +97,9 @@ describe('#getStepStartStates success', () => {
         disabled: false,
       },
       [STEP_IAM_SERVICE_ACCOUNTS]: {
+        disabled: false,
+      },
+      [STEP_RESOURCE_MANAGER_IAM_POLICY]: {
         disabled: false,
       },
     };