+++ src/index.ts
@@ -5,6 +5,7 @@ import { functionsSteps } from './steps/functions';
 import { storageSteps } from './steps/storage';
 import { serviceUsageSteps } from './steps/service-usage';
 import { iamSteps } from './steps/iam';
+import { resourceManagerSteps } from './steps/resource-manager';
 
 export const invocationConfig: IntegrationInvocationConfig<IntegrationConfig> = {
   instanceConfigFields: {
@@ -19,5 +20,6 @@ export const invocationConfig: IntegrationInvocationConfig<IntegrationConfig> = 
     ...storageSteps,
     ...serviceUsageSteps,
     ...iamSteps,
+    ...resourceManagerSteps,
   ],
 };