+++ src/steps/iam/converters.test.ts
@@ -1,64 +1,21 @@
-import { iam_v1 } from 'googleapis';
 import {
   createIamRoleEntity,
   createIamServiceAccountEntity,
   createIamServiceAccountKeyEntity,
   createIamServiceAccountHasKeyRelationship,
+  createIamUserEntity,
 } from './converters';
-
-function getMockIamRole(
-  partial?: Partial<iam_v1.Schema$Role>,
-): iam_v1.Schema$Role {
-  return {
-    name: 'projects/j1-gc-integration-dev/roles/myrole',
-    title: 'Role',
-    description: 'Description',
-    includedPermissions: [
-      'cloudfunctions.functions.get',
-      'cloudfunctions.functions.getIamPolicy',
-    ],
-    stage: 'GA',
-    etag: 'abc123',
-    ...partial,
-  };
-}
-
-function getMockServiceAccount(
-  partial?: Partial<iam_v1.Schema$ServiceAccount>,
-): iam_v1.Schema$ServiceAccount {
-  return {
-    name:
-      'projects/j1-gc-integration-dev/serviceAccounts/j1-gc-integration-dev-sa@j1-gc-integration-dev.iam.gserviceaccount.com',
-    projectId: 'j1-gc-integration-dev',
-    uniqueId: '1234567890',
-    email:
-      'j1-gc-integration-dev-sa@j1-gc-integration-dev.iam.gserviceaccount.com',
-    etag: 'abc=',
-    description: 'J1 Google Cloud integration execution',
-    oauth2ClientId: '1234567890',
-    ...partial,
-  };
-}
-
-function getMockServiceAccountKey(
-  partial?: Partial<iam_v1.Schema$ServiceAccountKey>,
-): iam_v1.Schema$ServiceAccountKey {
-  return {
-    name:
-      'projects/j1-gc-integration-dev/serviceAccounts/j1-gc-integration-dev-sa-tf@j1-gc-integration-dev.iam.gserviceaccount.com/keys/12345',
-    validAfterTime: '2020-08-05T18:05:19Z',
-    validBeforeTime: '2020-08-21T18:05:19Z',
-    keyAlgorithm: 'KEY_ALG_RSA_2048',
-    keyOrigin: 'GOOGLE_PROVIDED',
-    keyType: 'SYSTEM_MANAGED',
-    ...partial,
-  };
-}
+import {
+  getMockIamRole,
+  getMockServiceAccount,
+  getMockServiceAccountKey,
+} from '../../../test/mocks';
 
 describe('#createIamRoleEntity', () => {
   test('should convert custom role to entity', () => {
     expect(
       createIamRoleEntity(getMockIamRole(), {
+        projectId: 'abc123',
         custom: true,
       }),
     ).toMatchSnapshot();
@@ -67,6 +24,7 @@ describe('#createIamRoleEntity', () => {
   test('should convert noncustom role to entity', () => {
     expect(
       createIamRoleEntity(getMockIamRole(), {
+        projectId: 'abc123',
         custom: false,
       }),
     ).toMatchSnapshot();
@@ -120,6 +78,20 @@ describe('#createIamServiceAccountHasKeyRelationship', () => {
       createIamServiceAccountHasKeyRelationship({
         serviceAccountEntity,
         serviceAccountKeyEntity,
+      }),
+    ).toMatchSnapshot();
+  });
+});
+
+describe('#createIamUserEntity', () => {
+  test('should convert custom role to entity', () => {
+    expect(
+      createIamUserEntity({
+        projectId: 'abc123',
+        type: 'user',
+        identifier: 'test.user@example.com',
+        uniqueid: undefined,
+        deleted: false,
       }),
     ).toMatchSnapshot();
   });