+++ src/steps/iam/converters.ts
@@ -13,14 +13,28 @@ import {
   IAM_SERVICE_ACCOUNT_ENTITY_TYPE,
   IAM_SERVICE_ACCOUNT_KEY_ENTITY_CLASS,
   IAM_SERVICE_ACCOUNT_KEY_ENTITY_TYPE,
+  IAM_USER_ENTITY_CLASS,
+  IAM_USER_ENTITY_TYPE,
 } from './constants';
 import { generateEntityKey } from '../../utils/generateKeys';
+import { ParsedIamMemberType } from '../../utils/iam';
+
+export function toIamRoleEntityKey(params: {
+  projectId: string;
+  roleName: string;
+}): string {
+  // Managed roles do not contain the project ID. We just prefix all role
+  // entities with the project ID for safety.
+  return `project:${params.projectId}-role:${params.roleName}`;
+}
 
 export function createIamRoleEntity(
   data: iam_v1.Schema$Role,
   {
+    projectId,
     custom,
   }: {
+    projectId: string;
     /**
      * Google Cloud has managed roles and custom roles. There is no metadata
      * on the role itself that marks whether it's a custom role or managed role.
@@ -29,17 +43,19 @@ export function createIamRoleEntity(
     custom: boolean;
   },
 ) {
+  const roleName = data.name as string;
+
   return createIntegrationEntity({
     entityData: {
       source: data,
       assign: {
         _class: IAM_ROLE_ENTITY_CLASS,
         _type: IAM_ROLE_ENTITY_TYPE,
-        _key: generateEntityKey({
-          type: IAM_ROLE_ENTITY_TYPE,
-          id: data.name as string,
+        _key: toIamRoleEntityKey({
+          projectId,
+          roleName,
         }),
-        name: data.name,
+        name: roleName,
         displayName: data.title as string,
         description: data.description,
         stage: data.stage,
@@ -62,6 +78,43 @@ function getServiceAccountWebLink({
   return `https://console.cloud.google.com/iam-admin/serviceaccounts/details/${serviceAccountId}?orgonly=true&project=${projectId}&supportedpurview=organizationId`;
 }
 
+export function toIamUserEntityKey(params: {
+  projectId: string;
+  member: string;
+}): string {
+  // Managed roles do not contain the project ID. We just prefix all role
+  // entities with the project ID for safety.
+  return `project:${params.projectId}-user:${params.member}`;
+}
+
+export function createIamUserEntity(data: {
+  projectId: string;
+  type: ParsedIamMemberType;
+  identifier: string;
+  uniqueid: string | undefined;
+  deleted: boolean;
+}) {
+  return createIntegrationEntity({
+    entityData: {
+      source: data,
+      assign: {
+        _class: IAM_USER_ENTITY_CLASS,
+        _type: IAM_USER_ENTITY_TYPE,
+        _key: toIamUserEntityKey({
+          projectId: data.projectId,
+          member: data.identifier as string,
+        }),
+        displayName: data.identifier,
+        name: data.identifier,
+        username: data.identifier,
+        type: data.type,
+        deleted: data.deleted,
+        uniqueid: data.uniqueid,
+      },
+    },
+  });
+}
+
 export function createIamServiceAccountEntity(
   data: iam_v1.Schema$ServiceAccount,
 ) {
@@ -74,9 +127,9 @@ export function createIamServiceAccountEntity(
       assign: {
         _class: IAM_SERVICE_ACCOUNT_ENTITY_CLASS,
         _type: IAM_SERVICE_ACCOUNT_ENTITY_TYPE,
-        _key: generateEntityKey({
-          type: IAM_SERVICE_ACCOUNT_ENTITY_TYPE,
-          id: serviceAccountId,
+        _key: toIamUserEntityKey({
+          projectId,
+          member: data.email as string,
         }),
         name: data.name,
         displayName: (data.displayName || data.name) as string,