+++ src/steps/resource-manager/constants.ts
@@ -1,0 +1,7 @@
+export const STEP_RESOURCE_MANAGER_IAM_POLICY =
+  'fetch-resource-manager-iam-policy';
+
+export const IAM_SERVICE_ACCOUNT_IMPLEMENTS_ROLE_RELATIONSHIP_TYPE =
+  'google_cloud_iam_service_account_implements_role';
+export const IAM_USER_IMPLEMENTS_ROLE_RELATIONSHIP_TYPE =
+  'google_cloud_iam_user_implements_role';